+++ Dockerfile
@@ -2,8 +2,15 @@ FROM python:3.9-slim-buster
 
 WORKDIR /app
 
-COPY . .
-
+# Copy requirements first to leverage Docker cache
+COPY requirements.txt .
 RUN pip install -r requirements.txt
 
-CMD ["sh", "-c", "python scheduler.py & python processor.py & wait"]
+# Copy the rest of the application
+COPY . .
+
+# Create logs directory if it doesn't exist
+RUN mkdir -p logs
+
+# Run both scheduler and processor
+CMD ["sh", "-c", "python -m src.core.scheduler & python -m src.core.processor & wait"]
