+++ tests/test_processor.py
@@ -1,0 +1,10 @@
+import pytest
+from src.core import processor
+
+def test_check_mv_existence():
+    # Add tests here
+    pass
+
+def test_fetch_dependencies():
+    # Add tests here
+    pass
