+++ tests/test_scheduler.py
@@ -1,0 +1,10 @@
+import pytest
+from src.core import scheduler
+
+def test_build_dependency_graph():
+    # Add tests here
+    pass
+
+def test_collect_dependencies():
+    # Add tests here
+    pass
